Fix ByondUI small map preview (#90277)
This PR should fix the problem of small previews in TGUI once and for
all (I hope).

What was causing it? Because TGUI takes a long time to open, that's why
previews were generated broken (small).

On Byond 515 this problem was not so noticeable as the interfaces opened
faster, but with the release of 516 it became much worse.
Previews were generated inside a window that was not yet open, so the
scale was broken, sometimes the window would open before the preview was
done and sent, usually with small interfaces, or when reopening.

I'm not very good at working with signals, and to tell the truth this is
my second experience with them, so I hope I did it right.

No more small map previews

<details> <summary> Video </summary>

https://github.com/user-attachments/assets/834f3820-cc6a-4f65-90e5-d6bb2a118bcf

</details>

:cl:
fix: Fixed small character preview, color matrix preview, mech preview,
and other previews with uses ByondUI map
/:cl:

---------

Co-authored-by: Gaxeer <44334376+Gaxeer@users.noreply.github.com>