Tweaks rifle and sniper ammo (#4380)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Makes some tweaks to DMR and sniper ammo. Importantly, this allows .308
and .300 to break a couple damage thresholds.

- 8mm Illestren: 35->37 base damage, 40->30 AP
- 8mm Illestren Match: 30->27 AP.
- .308: 30->35 base damage
- .308 AP: 27->32 damage
- 8x58mm Caseless: 40->45 dam, 0->10 stamina, 45->50 AP
- .300: 45 -> 50 dam

## Why It's Good For The Game

.308 was the only DMR round with 4 shots to crit. 8mm Caseless was
almost completely unused and easily the worst of the three sniper
cartridges (tied for worst damage with no stamina side effect), on top
of the only weapon that uses it being exceedingly rare to begin with.
.300 goes to a 2 shot kill because it's .300 and the scout fires so
slowly that it still takes forever to make those 2 shots.

## Changelog

:cl:
balance: Buffed damage on .308, 8mm Illestren, 8x58mm Caseless, and
.300. Nerfed 8mm Illestren AP factor.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->