Nerfs buckshot AP, buffs .357 damage (#4347)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Nerfs buckshot to -30 AP, decaying to a max of -50 at long range. Buffs
.357 to deal 35 damage per shot, allowing it to triple tap unarmored
humanoids.

## Why It's Good For The Game

buckshot was a little too strong against armor again and was elbowing
out shotgun slugs at close range, .357 felt a little too anemic for what
it is

## Changelog

:cl:
balance: Nerfed buckshot, buffed .357
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->