Minor Riggs Changes (#4350)
## About The Pull Request

Makes a few changes to the Riggs:


![image](https://github.com/user-attachments/assets/dfc03dd9-f374-4561-bb00-6fd00499bd11)

- Expanded the cargo bay by one tile and shifted some things around in
there
- Re-decaled the medbay and added anesthetic
- Removed the flaming arrow, candor, and security belt from security.
Replaced with ESG, Sporter, and assault belt
- Added a Security Hardsuit for sec detail
- Ammo for all guns including Captain and First Mate
- Added a filter for the air harvester (it didnt have one before) along
with an indoor valve
- Plated catwalks for engineering
- Tiny bit of carpet for crew areas
- Gas mixer and a tiny bit extra fuel in port engine
- Atmospherics tank expanded by one tile, passive vent turned to
injector and unary vent so it can both inject and siphon

## Why It's Good For The Game

It's a pretty old ship. Tiny little touchup.

## Changelog

:cl:
add: Added new decaling and flooring in some areas to the Riggs, along
with an armament increase and expansion of the cargo bay.
add: Modified Riggs atmospherics and added anesthetic to its medbay
fix: Riggs airlock now works
/:cl:

---------

Co-authored-by: github-actions <action@github.com>