nerfs lucky #2 (#4444)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
turns it from a fully equiped dark gygax to just a wreck

Bernard is gonna be pissed. 

![image](https://github.com/user-attachments/assets/111d6a29-dfe6-4a0d-a417-549e2dcb1abc)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
its a little too strong for this ruin im told and we tack decent
incentives and systems(keeping the mech around locks up another mission
from spawning) to have people not just keep it.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
balance: the "lucky #2" mission is just a wreck and only worth 9k
fix: lucky #2 now has the proper index so it doesnt get mixed up with
the data retrieval mission
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->