Removes armor from boots, gloves, and jumpsuits (and a hat) (#4418)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Strips armor (except acid and fire) off of all non-adminspawn gloves,
boots, and jumpsuits that had any. And the PGF hat.

## Why It's Good For The Game

Gloves, jumpsuits, and boots combine their armor with anything worn over
the top that has the same coverage (such as a hardsuit). This probably
isn't good, especially since more powerful armor is more common here
than on a station server and we'll likely be adding more full-body suits
in the future.

## Changelog

:cl:
balance: Removed armor from gloves, boots, and jumpsuits. Armor is only
a feature of actual armor items and space suits.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->