Adding Smoke Grenades and Re-pricing Stingers in Cargo (#4449)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This adds smoke grenades to be cargo purchaseable in a pack of 7, much
like the flashbangs and teargas grenades, pricing the box at 500
credits.

This also adjusts the price of the stingbang grenades box from 1500 to
750 creating a more reasonable price per unit.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Smoke grenades already had a box but were unable to be purchased in
cargo. This simply updates their box description to something more
flavorful and makes the box accessible in cargo. The 500 credit price
tag puts the smoke grenades at about 71 credits per grenade, which I
believe is fair considering that smoke grenades can commonly spawn as
random loot on sand planets. They also lack any of the agressive
effects, such as those held by teargas grenades or flashbangs.

The price of stingbangs was lowered to 750 credits for the box of 5 as
to make them viable when compared with frag grenades. With the new
price, a single stingbang's cost is 150 credits, instead of the previous
300 credits per grenade. When compared to the 250 credit fragmentation
grenade, I think that this is a fair price change which may encourage
the use of stingers, or at least make them viable if nothing else.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added smoke grenades to cargo
balance: Rebalanced stinger grenade price
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: DIB-DOG <DaneOfThieves@gmail.com>
Co-authored-by: Changelogs <action@github.com>