Implementing a new Intermediate Etherbor Weapon (#4467)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds sprites and supporting code for a new Etherbor weapon, the BGC-10
beam carbine, intended to bridge the gap between the BG-16 and HBG
rifles.

The new BGC-10 uses Etherbor AR mode with 20 shots per cell and
performance comprable to 5.56. Notably, it lacks the DMR mode of the
HBG.

Accordingly, it has been added to the PGF factional market for the price
of 5000 credits, on par with other assault rifles of it's class, such as
the SKM and Hydra variants, as well as the CM-82.

Additionally, the price of the BG-16, which compares in performance to
the 9mm and 10mm submachineguns of other factions has had it's price
reduced from 3500 to 3000, bringing it in line with said weapons.

Finally, the maximum range of the Etherbor concentrated energy shots has
been adjusted from 10 to 12 tiles, and the Etherbor AR shots from 12 to
14 tiles.

Screenshots of sprites and inhands to follow:
![Screenshot 2025-03-23
041656](https://github.com/user-attachments/assets/ac78f5b5-a33c-4c07-8110-6f3ae3884863)
![Screenshot 2025-03-23
041701](https://github.com/user-attachments/assets/af9b1354-3add-41cc-b171-84fbf30e7e67)
![Screenshot 2025-03-23
041729](https://github.com/user-attachments/assets/2f50b92e-93f5-4e2a-9db8-2b19b1b28a11)
![Screenshot 2025-03-23
041732](https://github.com/user-attachments/assets/a3309689-1aa2-466c-b08a-5dc1bdaf4d5e)
![Screenshot 2025-03-23
041708](https://github.com/user-attachments/assets/600b5e6f-d91b-4c22-a7ab-968f66de5e1a)
![Screenshot 2025-03-23
041715](https://github.com/user-attachments/assets/025a067b-c544-4ef5-aca9-3b926d337266)
![Screenshot 2025-03-24
004321](https://github.com/user-attachments/assets/9449ae4e-3403-4cb2-85b6-7526c61d56a6)
![Screenshot 2025-03-24
004336](https://github.com/user-attachments/assets/a8a770d8-6438-4e3f-b795-7ed1a33ef510)
![Screenshot 2025-03-24
004349](https://github.com/user-attachments/assets/0e4b9676-e461-44ba-90c1-3dd2ccb14a81)

Bonus family photo:
![Screenshot 2025-03-23
042545](https://github.com/user-attachments/assets/6c45f3c3-7017-462f-9982-cf2beec2a2da)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The PGF as a faction has a very tight selection of weapons to choose
from. Given the fact that both HBG rifles are not availible in the cargo
market, for fear of their DMR mode being too powerful, that leaves the
PGF with what is effectively 2 guns. The SG-8 and the BG-12/16. The
addition of this intermediate weapon, aligns well with the ARs of other
factions and allows the PGF access to an assault rifle level weapon,
without the overcompensation of the HGB DMR rifles.

Making a max range change to Etherbor weapons isn't a decision I made
lightly, but there is good reason for it. Etherbor weapons are meant to
excel in close quarters, and weaken over distance, which this change is
still faithful to. Even with this change, at any range over 10 tiles,
these weapons do very little damage. This change, however, remedies the
issue of the beam's abrupt ending in outdoors or otherwise longer ranged
engagements. Hits scored within the added range will still only deal
nuisance levels of damage but will allow for the weapons to maintain a
presence in the fight where they were previously able to be wholly
written off.

Concerns over long range hitscan weapons are valid and in response,
etherbor weapons cannot be equipped with either of the scope attachment
variants. This leaves shots beyond the standard visible distance up to
guesswork and as previously stated, any shots _that are_ landed will be
severely neutered by the damage dropoff.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added the BGC-10 Beam Carbine
balance: rebalanced Etherbor weapon costs and performance
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: DIB-DOG <DaneOfThieves@gmail.com>
Co-authored-by: Changelogs <action@github.com>