Adds the KM-05 "Rottweiler" LMG (#4484)
## About The Pull Request

Adds a new Inteq firearm, a belt fed machine gun chambered in .308! It
is heavy and cumbersome, making it highly inaccurate and unwieldy when
fired without a bipod support. However, with the usage of a bipod, it is
accurate and stable when compared to other weapons of its class. The box
magazines have a 50 round capacity. The gun costs 6000 credits in cargo.


![rottweiler](https://github.com/user-attachments/assets/f7bfb2b2-4cf6-4aa8-ae19-8741f5fc9d4a)

## Why It's Good For The Game


https://github.com/user-attachments/assets/06a39a46-6e27-43e7-9746-60af282060bf


https://github.com/user-attachments/assets/3f636fee-0440-4459-9957-e59cfdbf50cf

## Changelog

:cl:
add: KM-05 Rottweiler machine gun to Inteq cargo catalogue
/:cl:

---------

Co-authored-by: Apogee-dev <60533805+Apogee-dev@users.noreply.github.com>