The Mercury-class Drua (SUNS exploration vessel) (#4395)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds in the Mercury class, a prospecting vessel used by SUNS to explore
the frontier and all its oddities. Decently equipped for exploring, also
capable of doing a nature documentary or two with its camera equipment.

Crew compliment is:
1 Captain
1 Academic (2ic)
2 Peacekeepers
1 Engineer
2 Students

images:

![image](https://github.com/user-attachments/assets/29f4c0c8-f0a1-4d62-9770-0f3561859e9d)

![image](https://github.com/user-attachments/assets/446aed34-5da5-46fa-981e-68d62d3dd0a4)

![image](https://github.com/user-attachments/assets/e1c23e21-e417-4ca0-8ee2-df39b0bf4614)


## Why It's Good For The Game
SUNS currently has one decently large ship, goal here is to give them a
smaller ship that should be easier to access for players looking for an
in to the faction. Hoping it makes them played again.

## Changelog

:cl:
add: Added ship

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->